+++ packages/common/src/allo/backends/allo-v2.ts
@@ -1,5 +1,5 @@
 import { Address, Hex } from "viem";
-import { Allo, AlloError, AlloOperation } from "../allo";
+import { Allo, AlloError, AlloOperation, CreateRoundArguments } from "../allo";
 import { error, Result, success } from "../common";
 import { WaitUntilIndexerSynced } from "../indexer";
 import { IpfsUploader } from "../ipfs";
@@ -16,7 +16,6 @@ import {
 } from "@allo-team/allo-v2-sdk/dist/types";
 import { Allo as AlloV2Contract, Registry } from "@allo-team/allo-v2-sdk/";
 import { AnyJson } from "../..";
-import { CreateRoundArguments } from "./allo-v1";
 
 export class AlloV2 implements Allo {
   private transactionSender: TransactionSender;
